hungarian: More performance tuning: Smaller datatypes are better for vectorized instructions

+++ src/hungarian.rs
@@ -32,13 +32,13 @@ pub type Score = u32;
 /// in the range 45001 -- 45101, so u16 is still sufficient. With 50 courses and max 20 places in each, the matrix will
 /// be 2MB in size, which is easily cachable.
 /// 
-/// For performance it seems to be better if EdgeWeight is the same type as Label, and even better when all are the same
-/// size as usize and bool. So we'll use isize instead of u16. This quadruples the size of the adjacency matrix in
-/// memory (on 64bit architectures), which is still cachable in L3 cache by modern CPUs.
-pub type EdgeWeight = isize;
+/// For performance it seems to be better if EdgeWeight is the same type as Label, and even better when edge weights and
+/// labels are the same size -- but smaller is better for vectorized instructions. So we'll use i32 instead of u16.
+/// This doubles the size of the adjacency matrix in memory, which is still cachable in L3 cache by modern CPUs.
+pub type EdgeWeight = i32;
 
-pub type Label = isize;
-const LARGE_LABEL: Label = std::isize::MAX;
+pub type Label = i32;
+const LARGE_LABEL: Label = std::i32::MAX;
 
 /// Execute the hungarian algorithm
 ///
@@ -372,7 +372,7 @@ mod tests {
         const NUM_COURSES: usize = 15;
         const PLACES_PER_COURSE: usize = 10;
         const NUM_PARTICIPANTS: usize = 100;
-        const WEIGHT_OFFSET: isize = 50000;
+        const WEIGHT_OFFSET: i32 = 50000;
         const CHOICES: usize = 3;
 
         let n = NUM_COURSES * PLACES_PER_COURSE;
